+++ namer/namer.cc
@@ -389,7 +389,7 @@ public:
         }
 
         if (klass->symbol != core::Symbols::root() && !klass->declLoc.file().data(ctx).isRBI() &&
-            classDefinesBehavior(klass)) {
+            ast::BehaviorHelpers::checkClassDefinesBehavior(klass)) {
             // TODO(dmitry) This won't find errors in fast-incremental mode.
             auto prevLoc = classBehaviorLocs.find(klass->symbol);
             if (prevLoc == classBehaviorLocs.end()) {
